Use xterm 256-color mode instead of 24 bit colors.

+++ include/graphics/AnsiPixel.h
@@ -6,12 +6,12 @@
 class AnsiPixel {
 private:
     char character;
-    uint32_t foreground;
-    uint32_t background;
+    uint8_t foreground;
+    uint8_t background;
 
 public:
     /// Construct a new AnsiPixel.
-    AnsiPixel(uint32_t background = Colors::NONE, uint32_t foreground = Colors::NONE,
+    AnsiPixel(uint8_t background = Colors::NONE, uint8_t foreground = Colors::NONE,
               char character = ' ') {
         this->character = character;
         this->foreground = foreground;
@@ -24,12 +24,12 @@ public:
     }
 
     /// Get the foreground color of this pixel.
-    inline uint32_t getForeground() {
+    inline uint8_t getForeground() {
         return foreground;
     }
 
     /// Get the background color of this pixel.
-    inline uint32_t getBackground() {
+    inline uint8_t getBackground() {
         return background;
     }
 };
